app/requirements.txt
twitchio			# Library Twitch
confluent-kafka		# Library for Kafka (Confluent)
kafka-python		# Kafka Python client